Finished tasks while hard coding the discount

+++ src/main/java/nz/ac/auckland/se281/Policy.java
@@ -8,7 +8,7 @@ public class Policy {
     this.sumToInsure = sumToInsure;
   }
 
-  public double basePremium() {
-    return 0.0;
+  public int basePremium() {
+    return 0;
   }
 }
